requirements.txt
# --- Airflow Core + Spark ---
apache-airflow==2.9.3
apache-airflow-providers-apache-spark==2.1.3
pyspark==3.5.2

# --- Airflow Providers ---
apache-airflow-providers-amazon
apache-airflow-providers-http
apache-airflow-providers-trino

# --- dbt ---
dbt-core==1.8.7
dbt-trino==1.8.4

# --- Optional / Utils ---
minio
requests
pandas
boto3
